serial rx synchonization
when the RX line is not idle at boot, the first reads are garbagge
until the receiver is properly synchonized.

With this by default the RX starts unsynchonized. All received
characters get dropped until it receives a MUOS_SERIAL_RXSYNC ['\r']
character. After (and including that character) the RX is synchronized and
received characters are pushed to the rx queue.

A Frame error will put the RX back into unsynchronized state.